refactored ai tracing to commonize types
