+++ .changeset/curvy-cycles-train.md
@@ -1,0 +1,5 @@
+---
+'@mastra/core': patch
+---
+
+"refactored ai tracing to commonize types"
