+++ packages/core/src/ai-tracing/index.ts
@@ -2,13 +2,13 @@
  * Mastra AI Tracing
  */
 
+
 // Core types and interfaces
-export * from './default';
+export * from './spans';
+export * from './tracers';
 export * from './types';
 export * from './registry';
 export * from './utils';
 export * from './context';
 export * from './exporters';
-
-// Abstract base class
-export { MastraAITracing } from './base';
+export * from './span_processors';
