+++ packages/core/src/ai-tracing/registry.ts
@@ -4,11 +4,17 @@
  * Provides a global registry for AI tracing instances.
  */
 
-import { MastraAITracing } from './base';
-import { DefaultAITracing, SensitiveDataFilter } from './default';
 import { CloudExporter, DefaultExporter } from './exporters';
+import { SensitiveDataFilter } from './span_processors';
+import { BaseAITracing, DefaultAITracing } from './tracers';
 import { SamplingStrategyType } from './types';
-import type { TracingSelector, AITracingSelectorContext, AITracingConfig, AITracingInstanceConfig } from './types';
+import type {
+  AITracing,
+  ConfigSelectorOptions,
+  ConfigSelector,
+  TracingConfig,
+  ObservabilityRegistryConfig,
+} from './types';
 
 // ============================================================================
 // Global AI Tracing Registry
@@ -18,14 +24,14 @@ import type { TracingSelector, AITracingSelectorContext, AITracingConfig, AITrac
  * Global registry for AI Tracing instances.
  */
 class AITracingRegistry {
-  private instances = new Map<string, MastraAITracing>();
-  private defaultInstance?: MastraAITracing;
-  private selector?: TracingSelector;
+  private instances = new Map<string, AITracing>();
+  private defaultInstance?: AITracing;
+  private configSelector?: ConfigSelector;
 
   /**
    * Register a tracing instance
    */
-  register(name: string, instance: MastraAITracing, isDefault = false): void {
+  register(name: string, instance: AITracing, isDefault = false): void {
     if (this.instances.has(name)) {
       throw new Error(`AI Tracing instance '${name}' already registered`);
     }
@@ -41,31 +47,31 @@ class AITracingRegistry {
   /**
    * Get a tracing instance by name
    */
-  get(name: string): MastraAITracing | undefined {
+  get(name: string): AITracing | undefined {
     return this.instances.get(name);
   }
 
   /**
    * Get the default tracing instance
    */
-  getDefault(): MastraAITracing | undefined {
+  getDefault(): AITracing | undefined {
     return this.defaultInstance;
   }
 
   /**
    * Set the tracing selector function
    */
-  setSelector(selector: TracingSelector): void {
-    this.selector = selector;
+  setSelector(selector: ConfigSelector): void {
+    this.configSelector = selector;
   }
 
   /**
    * Get the selected tracing instance based on context
    */
-  getSelected(context: AITracingSelectorContext): MastraAITracing | undefined {
+  getSelected(options: ConfigSelectorOptions): AITracing | undefined {
     // 1. Try selector function if provided
-    if (this.selector) {
-      const selected = this.selector(context, this.instances);
+    if (this.configSelector) {
+      const selected = this.configSelector(options, this.instances);
       if (selected && this.instances.has(selected)) {
         return this.instances.get(selected);
       }
@@ -98,13 +104,13 @@ class AITracingRegistry {
   clear(): void {
     this.instances.clear();
     this.defaultInstance = undefined;
-    this.selector = undefined;
+    this.configSelector = undefined;
   }
 
   /**
    * Get all registered instances
    */
-  getAll(): ReadonlyMap<string, MastraAITracing> {
+  getAll(): ReadonlyMap<string, AITracing> {
     return new Map(this.instances);
   }
 }
@@ -118,36 +124,36 @@ const aiTracingRegistry = new AITracingRegistry();
 /**
  * Register an AI tracing instance globally
  */
-export function registerAITracing(name: string, instance: MastraAITracing, isDefault = false): void {
+export function registerAITracing(name: string, instance: AITracing, isDefault = false): void {
   aiTracingRegistry.register(name, instance, isDefault);
 }
 
 /**
  * Get an AI tracing instance from the registry
  */
-export function getAITracing(name: string): MastraAITracing | undefined {
+export function getAITracing(name: string): AITracing | undefined {
   return aiTracingRegistry.get(name);
 }
 
 /**
  * Get the default AI tracing instance
  */
-export function getDefaultAITracing(): MastraAITracing | undefined {
+export function getDefaultAITracing(): AITracing | undefined {
   return aiTracingRegistry.getDefault();
 }
 
 /**
- * Set the AI tracing selector function
+ * Set the AI tracing config selector
  */
-export function setAITracingSelector(selector: TracingSelector): void {
+export function setSelector(selector: ConfigSelector): void {
   aiTracingRegistry.setSelector(selector);
 }
 
 /**
- * Get the selected AI tracing instance based on context
+ * Get the selected AI tracing instance based on options
  */
-export function getSelectedAITracing(context: AITracingSelectorContext): MastraAITracing | undefined {
-  return aiTracingRegistry.getSelected(context);
+export function getSelectedAITracing(options: ConfigSelectorOptions): AITracing | undefined {
+  return aiTracingRegistry.getSelected(options);
 }
 
 /**
@@ -174,7 +180,7 @@ export function clearAITracingRegistry(): void {
 /**
  * Get all registered AI tracing instances
  */
-export function getAllAITracing(): ReadonlyMap<string, MastraAITracing> {
+export function getAllAITracing(): ReadonlyMap<string, AITracing> {
   return aiTracingRegistry.getAll();
 }
 
@@ -193,16 +199,16 @@ export function hasAITracing(name: string): boolean {
 }
 
 /**
- * Type guard to check if an object is a MastraAITracing instance
+ * Type guard to check if an object is a AITracing instance
  */
-function isAITracingInstance(obj: Omit<AITracingInstanceConfig, 'name'> | MastraAITracing): obj is MastraAITracing {
-  return obj instanceof MastraAITracing;
+function isAITracingInstance(obj: Omit<TracingConfig, 'name'> | AITracing): obj is AITracing {
+  return obj instanceof BaseAITracing;
 }
 
 /**
- * Setup AI tracing from the AITracingConfig
+ * Setup AI tracing from the ObservabilityRegistryConfig
  */
-export function setupAITracing(config: AITracingConfig): void {
+export function setupAITracing(config: ObservabilityRegistryConfig): void {
   // Handle undefined/null config
   if (!config) {
     return;
@@ -247,6 +253,6 @@ export function setupAITracing(config: AITracingConfig): void {
 
   // Set selector function if provided
   if (config.configSelector) {
-    setAITracingSelector(config.configSelector);
+    setSelector(config.configSelector);
   }
 }
