+++ packages/core/src/ai-tracing/span_processors/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Mastra AI Span Processors
+ */
+
+export * from './sensitive-data-filter';
