+++ packages/core/src/ai-tracing/spans/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Mastra AI Spans
+ */
+
+export * from './base';
+export * from './default';
+export * from './no-op';
