+++ packages/core/src/ai-tracing/tracers/default.ts
@@ -1,0 +1,14 @@
+import { DefaultAISpan } from '../spans';
+import type { AISpanType, AISpan, TracingConfig, CreateSpanOptions } from '../types';
+import { BaseAITracing } from './base';
+
+export class DefaultAITracing extends BaseAITracing {
+  constructor(config: TracingConfig) {
+    super(config);
+  }
+
+  protected createSpan<TType extends AISpanType>(options: CreateSpanOptions<TType>): AISpan<TType> {
+    // Simple span creation - base class handles all tracing lifecycle automatically
+    return new DefaultAISpan<TType>(options, this);
+  }
+}
