+++ packages/core/src/ai-tracing/utils.ts
@@ -3,80 +3,9 @@
  * used in AI tracing and observability.
  */
 
-import type { RuntimeContext } from '../di';
+import type { RuntimeContext } from '../runtime-context';
 import { getSelectedAITracing } from './registry';
 import type { AISpan, AISpanType, AISpanTypeMap, AnyAISpan, TracingContext, TracingOptions } from './types';
-
-const DEFAULT_KEYS_TO_STRIP = new Set([
-  'logger',
-  'experimental_providerMetadata',
-  'providerMetadata',
-  'steps',
-  'tracingContext',
-]);
-export interface DeepCleanOptions {
-  keysToStrip?: Set<string>;
-  maxDepth?: number;
-}
-
-/**
- * Recursively cleans a value by removing circular references and stripping problematic or sensitive keys.
- * Circular references are replaced with "[Circular]". Unserializable values are replaced with error messages.
- * Keys like "logger" and "tracingContext" are stripped by default.
- * A maximum recursion depth is enforced to avoid stack overflow or excessive memory usage.
- *
- * @param value - The value to clean (object, array, primitive, etc.)
- * @param options - Optional configuration:
- *   - keysToStrip: Set of keys to remove from objects (default: logger, tracingContext)
- *   - maxDepth: Maximum recursion depth before values are replaced with "[MaxDepth]" (default: 10)
- * @returns A cleaned version of the input with circular references, specified keys, and overly deep values handled
- */
-export function deepClean(
-  value: any,
-  options: DeepCleanOptions = {},
-  _seen: WeakSet<any> = new WeakSet(),
-  _depth: number = 0,
-): any {
-  const { keysToStrip = DEFAULT_KEYS_TO_STRIP, maxDepth = 10 } = options;
-
-  if (_depth > maxDepth) {
-    return '[MaxDepth]';
-  }
-
-  if (value === null || typeof value !== 'object') {
-    try {
-      JSON.stringify(value);
-      return value;
-    } catch (error) {
-      return `[${error instanceof Error ? error.message : String(error)}]`;
-    }
-  }
-
-  if (_seen.has(value)) {
-    return '[Circular]';
-  }
-
-  _seen.add(value);
-
-  if (Array.isArray(value)) {
-    return value.map(item => deepClean(item, options, _seen, _depth + 1));
-  }
-
-  const cleaned: Record<string, any> = {};
-  for (const [key, val] of Object.entries(value)) {
-    if (keysToStrip.has(key)) {
-      continue;
-    }
-
-    try {
-      cleaned[key] = deepClean(val, options, _seen, _depth + 1);
-    } catch (error) {
-      cleaned[key] = `[${error instanceof Error ? error.message : String(error)}]`;
-    }
-  }
-
-  return cleaned;
-}
 
 /**
  * Removes specific keys from an object.
@@ -148,6 +77,21 @@ function setNestedValue(obj: any, path: string, value: any): void {
 }
 
 /**
+ * Extracts the trace ID from a span if it is valid.
+ *
+ * This helper is typically used to safely retrieve the `traceId` from a span object,
+ * while gracefully handling invalid spans — such as no-op spans — by returning `undefined`.
+ *
+ * A span is considered valid if `span.isValid` is `true`.
+ *
+ * @param span - The span object to extract the trace ID from. May be `undefined`.
+ * @returns The `traceId` if the span is valid, otherwise `undefined`.
+ */
+export function getValidTraceId(span?: AnyAISpan): string | undefined {
+  return span?.isValid ? span.traceId : undefined;
+}
+
+/**
  * Creates or gets a child span from existing tracing context or starts a new trace.
  * This helper consolidates the common pattern of creating spans that can either be:
  * 1. Children of an existing span (when tracingContext.currentSpan exists)
@@ -188,28 +132,14 @@ export function getOrCreateSpan<T extends AISpanType>(options: {
     runtimeContext: runtimeContext,
   });
 
-  return aiTracing?.startSpan({
+  return aiTracing?.startSpan<T>({
     type,
     attributes,
-    startOptions: {
+    customSamplerOptions: {
       runtimeContext,
+      metadata,
     },
     ...rest,
     metadata,
   });
-}
-
-/**
- * Extracts the trace ID from a span if it is valid.
- *
- * This helper is typically used to safely retrieve the `traceId` from a span object,
- * while gracefully handling invalid spans — such as no-op spans — by returning `undefined`.
- *
- * A span is considered valid if `span.isValid` is `true`.
- *
- * @param span - The span object to extract the trace ID from. May be `undefined`.
- * @returns The `traceId` if the span is valid, otherwise `undefined`.
- */
-export function getValidTraceId(span?: AnyAISpan): string | undefined {
-  return span?.isValid ? span.traceId : undefined;
 }
